+++ frontend/src/components/lastBuildTable/BuildInfoTable.jsx
