+++ frontend/src/components/lastBuildTable/FailuresTable.jsx
